test: 포인트 타입 수정 (Long -> Int)

+++ src/test/kotlin/com/marketboro/controller/helper/TestRequestDto.kt
@@ -2,5 +2,5 @@ package com.marketboro.controller.helper
 
 
 data class TestPointTransactionReq(
-    val points: Long
+    val points: Int
 )
